test/e2e/requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
boto3==1.16.35
botocore==1.19.35
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
google-auth==1.24.0
google-pasta==0.2.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
isort==5.6.4
jmespath==0.10.0
kubernetes==12.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pluggy==0.13.1
protobuf==3.14.0
protobuf3-to-dict==0.1.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
smdebug-rulesconfig==1.0.0
toml==0.10.2
urllib3==1.26.2
websocket-client==0.57.0
wrapt==1.12.1
zipp==3.4.0
